+++ packages/stream_feeds/test/test_utils/event_types.dart
@@ -1,4 +1,6 @@
 class EventTypes {
+  static const String activityMarked = 'feeds.activity.marked';
+
   static const String followCreated = 'feeds.follow.created';
   static const String followDeleted = 'feeds.follow.deleted';
   static const String followUpdated = 'feeds.follow.updated';
@@ -7,4 +9,6 @@ class EventTypes {
   static const String pollDeleted = 'feeds.poll.deleted';
   static const String pollVoteCasted = 'feeds.poll.vote_casted';
   static const String pollVoteRemoved = 'feeds.poll.vote_removed';
+
+  static const String storiesFeedUpdated = 'feeds.stories_feed.updated';
 }